Move merge_prefect_task_configs to utils.py
--- don't change below this line ---
MLOPS-287 <#DTT#>
