+++ omigami/spec2vec/tasks/make_embeddings.py
@@ -7,7 +7,7 @@ from omigami.data_gateway import SpectrumDataGateway
 from omigami.gateways.redis_spectrum_gateway import REDIS_DB
 from omigami.spec2vec.helper_classes.embedding_maker import EmbeddingMaker
 from omigami.spec2vec.helper_classes.progress_logger import TaskProgressLogger
-from omigami.config import merge_prefect_task_configs
+from omigami.utils import merge_prefect_task_configs
 
 
 class MakeEmbeddings(Task):
