+++ omigami/spec2vec/tasks/process_spectrum/task.py
@@ -6,7 +6,7 @@ from prefect import Task
 from omigami.data_gateway import SpectrumDataGateway, InputDataGateway
 from omigami.gateways.redis_spectrum_gateway import REDIS_DB
 from omigami.spec2vec.helper_classes.progress_logger import TaskProgressLogger
-from omigami.config import merge_prefect_task_configs
+from omigami.utils import merge_prefect_task_configs
 from omigami.spec2vec.tasks.process_spectrum.spectrum_processor import (
     SpectrumProcessor,
 )
