+++ omigami/spec2vec/tasks/train_model.py
@@ -13,7 +13,7 @@ from omigami.data_gateway import SpectrumDataGateway
 from omigami.spec2vec.helper_classes.train_logger import (
     CustomTrainingProgressLogger,
 )
-from omigami.config import merge_prefect_task_configs
+from omigami.utils import merge_prefect_task_configs
 
 
 @dataclass
